find triangle

+++ codility/src/main/java/self/codility/sort/Triangle.java
@@ -8,7 +8,7 @@ public class Triangle {
         int retVal = 0;
         Arrays.sort(A);
         for (int index = A.length - 1; index >= 2; index--) {
-            long tmp = A[index - 2] + A[index - 1];
+            long tmp = (long)A[index - 2] + (long) A[index - 1];
             if (tmp > A[index]) {
                 retVal = 1;
                 break;
@@ -16,4 +16,5 @@ public class Triangle {
         }
         return retVal;
     }
+
 }
